refactor : article page with prefetchArticleQuery
