+++ src/app/article/[articleId]/page.tsx
@@ -1,52 +1,38 @@
-import { FewResponse } from "@api/fewFetch";
-import queryClient from "@api/queryClient";
 import ArticleTitle from "@article/components/ArticleTitle";
 import EmailContentTemplate from "@article/components/EmailContentTemplate";
-import { getArticleQueryOptions } from "@article/remotes/getArticleQueryOptions";
-import { getArticleWithWorkbookQueryOptions } from "@article/remotes/getArticleWithWorkbookQueryOptions";
-import { ArticleDetail } from "@article/types";
+import { prefetchArticleQuery } from "@article/remotes/prefetchArticleQuery";
+import { ArticlePageProps } from "@article/types";
 import { createMetadata } from "@shared/utils/metadata";
+import { HydrationBoundary } from "@tanstack/react-query";
 import { Metadata } from "next";
 
 export async function generateMetadata({
   params,
   searchParams,
-}: {
-  params: { articleId: string };
-  searchParams: { [key: string]: string | string[] | undefined };
-}): Promise<Metadata> {
-  const articleId = params.articleId;
-  const workbookId = searchParams?.workbookId;
+}: ArticlePageProps): Promise<Metadata> {
+  const { data: articleInfo } = await prefetchArticleQuery({
+    params,
+    searchParams,
+  });
 
-  let articleInfo = {} as FewResponse<ArticleDetail>;
-  if (typeof workbookId === "string") {
-    const { data } = await queryClient.fetchQuery({
-      ...getArticleWithWorkbookQueryOptions({
-        workbookId,
-        articleId,
-      }),
-    });
-
-    articleInfo = data;
-  } else {
-    const { data } = await queryClient.fetchQuery({
-      ...getArticleQueryOptions({ articleId }),
-    });
-    articleInfo = data;
-  }
-
-  const { title, writer } = articleInfo.data;
+  const { title, writer } = articleInfo;
   return createMetadata({
     title: title,
     description: `${writer.name} 작가의 ${title} 아티클 입니다.`,
   });
 }
+export default async function ArticlePage({
+  params,
+  searchParams,
+}: ArticlePageProps) {
+  const { state } = await prefetchArticleQuery({ params, searchParams });
 
-export default function ArticlePage() {
   return (
-    <div className="flex h-full flex-col gap-[46px]">
-      <ArticleTitle />
-      <EmailContentTemplate />
-    </div>
+    <HydrationBoundary state={state}>
+      <div className="flex h-full flex-col gap-[46px]">
+        <ArticleTitle />
+        <EmailContentTemplate />
+      </div>
+    </HydrationBoundary>
   );
 }
